requirements.txt
streamlit
openai
requests
beautifulsoup4
duckduckgo-search
wikipedia
pyyaml
typing